backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pymodbus==3.6.2
pyserial==3.5
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
python-multipart==0.0.6
pyyaml==6.0.1
aiofiles==23.2.1

